Subclass the db class and rename it to DBfile
